requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.26.25
botocore==1.29.25
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.3